pack/pyproject.toml
[project]
name = "pack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.6",
    "pact-python>=2.3.0",
    "pytest>=8.3.4",
    "requests>=2.32.3",
    "uvicorn>=0.34.0",
]
